correct

+++ Assembly-CSharp/ProtoBuf/Meta/MetaType.cs
@@ -1727,7 +1727,7 @@ namespace ProtoBuf.Meta
 			}
 			else
 			{
-				this.flags &= ~flag;
+				this.flags &= (byte) ~flag;
 			}
 		}
